Tests/LibWeb/Text/expected/Editing/execCommand-justifyLeft.txt

foobar
<div style="">foobar</div>
